+++ benchmarks/loop.go
@@ -1,12 +1,28 @@
 package main
 
-import "fmt"
+import (
+    "fmt"
+    "os"
+    "strconv"
+    "time"
+)
 
 func main() {
-    var sum int = 0
-    var max int = 100000000
-    for i := 0; i < max; i++ {
-        sum += i
+    t := int64(time.Now().Nanosecond())
+
+    n := uint64(100000000)
+    if len(os.Args) > 1 {
+        if v, err := strconv.ParseUint(os.Args[1], 10, 64); err == nil {
+            n = v
+        }
     }
-    fmt.Printf("Sum = %d\n", sum)
+
+    var acc uint64
+    for i := uint64(1); i <= n; i++ {
+        acc += i
+        if t*int64(i) < 0 {
+            acc = 0
+        }
+    }
+    fmt.Println(acc)
 }